requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
google-generativeai==0.4.0
Pillow==10.2.0
pdf2image==1.16.3
python-dotenv==1.0.0
selenium==4.17.2
webdriver-manager==4.0.1